refactored location of random walk
