+++ libp2p/discovery/random_walk/exceptions.py
